Changed events that triggered join/leave messages
No longer displays messages for players who are kicked while joining the
default server.  This fixes showing join/leave messages for players who
are banned on the default server, when a Bungee plugin isn't being used
to handle bans.  Uses a bit more RAM to keep a list of players, this
plugin probably isn't going to be used on larger networks anyhow.